services/auth/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.2
python-dotenv==1.0.0
PyJWT==2.8.0
httpx==0.27.0
